Issue #431: Working on the bl-backend-list component and all of the pieces needed to get the general architecture working. - Changed around the naming of the swagger endpoints so that the tags are the list names and the operation ids are action verbs with the list names after - Added `status` property to the view mixin along with `activate` and `deactivate` methods. - Added code to `<bl-app>` to activate the home view on startup. This hardcoding is temporary. Once there are multiple pages we will transition the activation over to the `<app-route>` based routing logic. - Built the bl-backend-list component with the basic architecture (notes about that below). It interacts with the parent view and the app-backend-client to make sure the api isn't touched until the view is active and the client is initialized. - Added groups and badges lists to `<bl-view-home>`.
**Notes on the bl-backend-list architecture:**
It is loosely based on the FLUX design pattern from facebook / react.
The idea is that the bl-backend-list (and eventually the bl-backend-item) components act as the dispatchers.
They do this by adding callbacks to the binding targets in the view components (groups and badges in this initial example).
Each view is ultimately in control of when all of its child backend components are free to communicate with the server.
So in that way the data flow is unidirectional down the DOM tree (from bl-view-* and bl-backend-* down to the child components).
But instead of using events to bubble actions back up the tree we use the callbacks right on the end objects (which simplifies the connection and also has the benefit of preventing evesdropping along the event bubbling path which would break the FLUX pattern).