Issue #523: Restructured the api docs again. This time to simplify everything.
It was getting a bit out of hand with my previous strategy of creating a different set of swagger docs for each different user type.

The new model only has two apis: internal and external. The complexity within the external API is all handled by explaining the permissions model in the markdown. I've included a new permissions table that is dynamically generated based on the application policy.

I'm hoping that this makes it less painful to make additions to the api going forward.