Issue #523: Refactored to resolve issues with new app model.
Issue 1:

The app model previously had relationships pointing to the active users / pending users / etc (and the same for groups). These resulted in the storage of giant lists of user and groups ids on the app. In production, while running the rake script to link all 30k existing users with the badge list standard app this proved impossible to manage. The process bugged out once it hit 9k users.

This was essentially a wake up call to the impact of has_and_belongs_to_many relationships on the database. Those sorts of relationships are best modeled as their own linker relationship (which they were in this case, but which they currently are not for group user members... that will be resolved soon). And when building a "cache" of the ids one the linked models (app ids on the user / user ids on the app) in order to save queries, basically the cached ids should be a manually created array of bson ids, not a has_and_belongs_to_many relationship and should exist only if the hosting model is unlikely to have more than 100-500 or so linked items. (So in this example it's ok to keep a list of active/pending/etc apps on the user since the list of app ids will never get unmanagably large. But the list of user ids on the app is fundamentally problematic and must be avoided.)

The resolution to this issue was to refactor the relationships between apps and users/groups to remove the app side of the relationship. While doing this I also removed all of the shortcut routes groups/id/apps, apps/id/groups, etc... Going forward these sorts of "crawl-through" indexes should be avoided because they are confusing and not real... they are better accomplished by an index on the root of the linker object which is able to embed the other parent object... that way the filter/sort values are more logical (they match up with the main object being returned) and you end up with the linker object and the both parent objects full details in a way that is semanticaly predictable and aligned with JSON API rb. I will need to refactor the groups and badges indexes to match up with this once the linker objects are created.

Issue 2:

There were extra queries occuring on all of the index routes do to the policy index and the json api renderer both querying the db for the list.

I refactored all of the api index controllers to use the '.entries' method to resolve the query before padding it to the rendering line. This saved from 1 to 3 queries per request (depending on the depth of the resource).