Issue #563: Changed what shows up on the badge page for private badges. This was done per a request from SAE, but after some thought I've decided to make the changes universal (rather than having it be a new badge setting).
Really the contents of the badge itself have always been public. My assumption at this point is that most people are using the badge visibility in order to obscure the identities of the badge members rather than obscuring the details of the badge. Also when all of the details of the badge are hidden it makes the badge much less valuable to the people who have earned it.

Since this represents a slight tweaking of the meaning of "private" and "secret" with respect to badge visibility, I have updated the descriptive text in the badge settings dialog.